==> 05_assessing.ipynb <==
# Avaliando
Use o espaço abaixo para explorar os arquivos `all_alpha_08.csv` e `all_alpha_18.csv` para responder as perguntas do teste abaixo.

import pandas as pd

# Data from 2008

df_08 = pd.read_csv('data/all_alpha_08.csv')
df_08.info()

df_08.shape

# Data from 2018

df_18 = pd.read_csv('data/all_alpha_18.csv')
df_18.info()

df_18.shape

df_18.duplicated().sum()

df_18.isnull().any(axis=1).sum()

df_18[df_18['SmartWay'].notnull()]['SmartWay'].nunique()

df_18[df_18['Cert Region'].notnull()]['Cert Region'].nunique()

df_08[df_08['Sales Area'].notnull()]['Sales Area'].nunique()

df_08[df_08['Trans'].notnull()]['Trans'].nunique()

df_08.head()

df_18.head()

df_08['Cyl'].nunique()

df_18['Cyl'].nunique()

df_08['Fuel'].unique()

df_18['Fuel'].unique()
