==> 06_cleaning_column_labels.ipynb <==
# Limpando rótulos das colunas
Use os arquivos `all_alpha_08.csv` e `all_alpha_18.csv`

# carregue os conjuntos de dados
import pandas as pd
df_08 = pd.read_csv('data/all_alpha_08.csv')
df_18 = pd.read_csv('data/all_alpha_18.csv')

# exiba o conjunto de dados de 2008
df_08.head(1)

# exiba o conjunto de dados de 2018
df_18.head(1)

### Descarte colunas irrelevantes

# descarte colunas do conjunto de dados de 2008
df_08.drop(['Stnd', 'Underhood ID', 'FE Calc Appr', 'Unadj Cmb MPG'], axis=1, inplace=True)

# confirme as mudanças
df_08.head(1)

# descarte colunas do conjunto de dados de 2018
df_18.drop(['Stnd', 'Stnd Description', 'Underhood ID', 'Comb CO2'], axis=1, inplace=True)

# confirme as mudanças
df_18.head(1)

### Renomeie as colunas

df_08.columns

# renomeie Sales Area para Cert Region
df_08.rename(columns={'Sales Area': 'Cert Region'}, inplace=True)

# confirme as mudanças
df_08.head(1)

# substitua espaços por rótulos com subtrações e letras minúsculas para o conjunto de dados de 2008
df_08.rename(columns=lambda x: x.strip().lower().replace(" ", "_"), inplace=True)

# confirme as mudanças
df_08.head(1)

# substitua espaços por rótulos com subtrações e letras minúsculas para o conjunto de dados de 2018
df_18.rename(columns=lambda label: label.strip().lower().replace(" ", "_"), inplace=True)

# confirme as mudanças
df_18.head(1)

# verifique se os rótulos das colunas dos conjuntos de dados de 2008 e 2018 são idênticos
df_08.columns == df_18.columns

# certifique-se que eles são todos idênticos assim
(df_08.columns == df_18.columns).all()

# salve os novos conjuntos de dados para a próxima seção
df_08.to_csv('data/data_08.csv', index=False)
df_18.to_csv('data/data_18.csv', index=False)
