==> 07_query_filter.ipynb <==
# Filtre, descarte valores nulos, desduplique
Use os arquivos `data_08.csv` e `data_18.csv`

# carregue conjuntos de dados
import pandas as pd
df_08 = pd.read_csv('data/data_08.csv')
df_18 = pd.read_csv('data/data_18.csv')

# exiba as dimensões do conjunto de dados
df_08.shape

# exiba as dimensões do conjunto de dados
df_18.shape

## Filtre por região de certificação

df_08.query('cert_region == "CA"')

# filtre conjuntos de dados por linhas segundo os padrões da Califórnia
df_08 = df_08.query("cert_region == 'CA'")
df_18 = df_18.query("cert_region == 'CA'")

# confirme que a única região de certificação é a Califórnia
df_08['cert_region'].unique()

# confirme que a única região de certificação é a Califórnia
df_18['cert_region'].unique()

# descarte a coluna de certificação de região de ambos conjuntos de dados
df_08.drop('cert_region', axis=1, inplace=True)
df_18.drop('cert_region', axis=1, inplace=True)


df_08.shape

df_18.shape

## Descarte linhas com valores faltantes

# exiba a contagem de valores faltantes para cada característica em 2008
df_08.isnull().sum()

# exiba a contagem de valores faltantes para cada característica em 2018
df_18.isnull().sum()

# descarte linhas com quaisquer valores nulos em ambos conjuntos de dados
df_08.dropna(axis=0, how='any', inplace=True)
df_08.shape

df_18.dropna(axis=0, how='any', inplace=True)
df_18.shape

# verifique se quaisquer colunas em 2008 têm valores nulos - o resultado deve ser False
df_08.isnull().sum().any()

# verifique se quaisquer colunas em 2018 têm valores nulos - o resultado deve ser False
df_18.isnull().sum().any()

## Desduplique os dados

# exiba o número de duplicatas nos conjuntos de dados de 2008 e 2018
df_08.duplicated().sum()

df_18.duplicated().sum()

# descarte duplicatas em ambos conjuntos de dados
df_08.drop_duplicates(keep='first', inplace=True)
df_18.drop_duplicates(keep='first', inplace=True)

# exiba o número de duplicatas novamente para confirmar desduplicação - ambos devem ser 0
print('08: ', df_08.duplicated().sum())
print('18: ', df_18.duplicated().sum())


# salve seu progresso para a próxima seção
df_08.to_csv('data/data_08.csv', index=False)
df_18.to_csv('data/data_18.csv', index=False)
