==> 09_fix_datatypes_cyl.ipynb <==
# Corrija o tipo de dados de `cyl`
- 2008: extraia int de string
- 2018: converta de float para int

# carregue os conjuntos de dados
import pandas as pd
df_08 = pd.read_csv('data/data_08.csv')
df_18 = pd.read_csv('data/data_18.csv')

# verifique a contagem de valores para a coluna cyl de 2008
df_08['cyl'].value_counts()

Leia [isto](https://stackoverflow.com/questions/35376387/extract-int-from-string-in-pandas) para te ajudar a extrair ints de strings no Pandas para o próximo passo.

# Extraia int de strings da coluna cyl de 2008
df_08['cyl'] = df_08['cyl'].str.extract('(\d+)', expand=False).astype(int)

# Verifique a contagem de valores para a coluna cyl de 2008 novamente para confirmar a mudança
df_08['cyl'].value_counts()

# converta a coluna cyl de 2018 para int
df_18['cyl'] = df_18['cyl'].astype(int)

df_08.to_csv('data/data_08.csv', index=False)
df_18.to_csv('data/data_18.csv', index=False)

df_08.shape

df_18.shape
