==> 13_drawing_conclusions.ipynb <==
# Tirando conclusões
Use o espaço abaixo para resolver questões sobre os conjuntos de dados `clean_08.csv` e `clean_18.csv`

# carregue os conjuntos de dados
import pandas as pd
import matplotlib as plt
%matplotlib inline

df_08 = pd.read_csv('data/clean_08.csv')
df_18 = pd.read_csv('data/clean_18.csv')

### Q1: Existem mais modelos únicos usando fontes alternativas de combustível? Em qual proporção?

df_08['smartway'] = df_08['smartway'].apply(lambda x: x.lower())
df_18['smartway'] = df_18['smartway'].apply(lambda x: x.lower())

#### Número de veículos 2008 usando fontes alternativas de combustível

smart_08 = df_08[df_08['smartway'] != 'no']['model'].nunique()

perc_smart_08 = smart_08 / df_08['model'].nunique()

print('Modelos únicos de 2008: {}'.format(str(df_08['model'].nunique())))
print('Percentual de modelos smartway de 2008: {:.3}'.format(perc_smart_08))

#### Número de veículos 2018 usando fontes alternativas de combustível

smart_18 = df_18[df_18['smartway'] != 'no']['model'].nunique()

perc_smart_18 = smart_18 / df_18['model'].nunique()

print('Modelos únicos de 2018: {}'.format(str(df_18['model'].nunique())))
print('Percentual de modelos smartway de 2018: {:.3}'.format(perc_smart_18))

perc_smart_08 - perc_smart_18

#### Q1, não, houve uma queda no percentual de carros usando fontes alternativas de combustível

### Q2: Qual foi o nível de melhoria das classes de veículo no quesito "economia de combustível"?

df_08[['city_mpg', 'hwy_mpg', 'cmb_mpg']].mean().describe()

df_18[['city_mpg', 'hwy_mpg', 'cmb_mpg']].mean().describe()

df_18[['city_mpg', 'hwy_mpg', 'cmb_mpg']].mean().describe()['mean'] / df_08[['city_mpg', 'hwy_mpg', 'cmb_mpg']].mean().describe()['mean']

#### Q2. Houve uma média de 20% de melhoria na economia de combustível.

### Q3: Quais são as características dos veículos SmartWay? Elas mudaram com o tempo?

df_08.query('smartway != "no"').describe()

df_18.query('smartway != "no"').describe()

#### Q3. A classificação SmartWay de um veículo é determinada pela relação entre os valoes de MPG e GreenHouse Gas Score. As referências são atualizadas para que as metas possam ser revisadas e motores otimizados.

### Q4: Quais características estão associadas à uma maior economia de combustível?

#### Q4. Os valores indicados em MPG, para City, Highway e CMB (Combined City and Highway), que representam a distância percorrida com um galão de combustível.
